requirements.txt
python-telegram-bot==20.7
flask
requests
pandas
numpy
nest_asyncio
